requirements.txt
flask
flask-cors
requests
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2